move time formatter to impl Display for speedrun timer
chore: fmt/clippy
