+++ src/util/mod.rs
@@ -1,5 +1,2 @@
 // Math
 pub mod quaternion;
-
-// Timer
-pub mod time_formatter;
